♻️ [Refactoring]: #433 [FE] TreeView の TreeNodeItem を独立ファイルへ抽出
TreeView/index.tsx 内に定義していた再帰ノード TreeNodeItem を
TreeNodeItem/index.tsx へ切り出し、TreeView からは import で参照する。
挙動・DOM 出力・公開 API・呼び出し元（BoardWorkspace）は不変の純内部リファクタ。

Refs #433
